MODIFY: Feed, Nav 컴포넌트화
