+++ public/data/commentData.json
@@ -66,13 +66,13 @@
     "comment": "Nice weather, nice timing"
   },
   {
-    "id": 10,
+    "id": 11,
     "feedId": 9,
     "username": "Bradjanet",
     "comment": "wow lucky you!"
   },
   {
-    "id": 11,
+    "id": 12,
     "feedId": 10,
     "username": "tomorrow",
     "comment": "can't wait to see it in person"
